Check if source files are present

+++ install.sh
@@ -1,6 +1,10 @@
 #!/bin/zsh
 
 install_file() {
+	if [ ! -f $1 ]; then
+		echo "$1 not found!" >&2
+		exit 1
+	fi
 	if [ -f $2 ]; then
 		echo "$2 is present."
 		rm $2
